Rearrange kubelet networking codes
