+++ pkg/kubelet/BUILD
@@ -18,7 +18,6 @@ go_library(
         "kubelet_pods.go",
         "kubelet_resources.go",
         "kubelet_volumes.go",
-        "networks.go",
         "oom_watcher.go",
         "pod_container_deletor.go",
         "pod_workers.go",
@@ -153,7 +152,6 @@ go_test(
         "kubelet_resources_test.go",
         "kubelet_test.go",
         "kubelet_volumes_test.go",
-        "networks_test.go",
         "oom_watcher_test.go",
         "pod_container_deletor_test.go",
         "pod_workers_test.go",
